Refactor top level library structure
Refactor top level library structure
