+++ src/__tests__/byte-stream.test.ts
@@ -1,5 +1,5 @@
 // @ts-nocheck
-import { generateSecureToken } from '../index';
+import { generateSecureToken } from '../generateSecureToken';
 import { randomBytes } from 'crypto-browserify';
 
 /* eslint-disable max-len */
