+++ src/__tests__/skip-padding.test.ts
@@ -1,11 +1,13 @@
 // @ts-nocheck
-import { gen } from '../index';
+import { generateSecureToken } from '../generateSecureToken';
 
 test('setting skipPadding to true may result in varied string token lengths if length is set to >1', () => {
     const desiredTokenLength = 6;
     let smallerTokenFound = false;
     while (!smallerTokenFound) {
-        const token = gen(desiredTokenLength, { skipPadding: true });
+        const token = generateSecureToken(desiredTokenLength, {
+            skipPadding: true,
+        });
         smallerTokenFound = token.length < desiredTokenLength;
     }
 });
@@ -14,7 +16,7 @@ test('setting skipPadding to true may result in varied number token lengths if l
     const desiredTokenLength = 6;
     let smallerTokenFound = false;
     while (!smallerTokenFound) {
-        const token = gen(desiredTokenLength, {
+        const token = generateSecureToken(desiredTokenLength, {
             skipPadding: true,
             returnType: 'number',
         });
@@ -26,7 +28,7 @@ test('setting skipPadding to true may result in varied bigint token lengths if l
     const desiredTokenLength = 6;
     let smallerTokenFound = false;
     while (!smallerTokenFound) {
-        const token = gen(desiredTokenLength, {
+        const token = generateSecureToken(desiredTokenLength, {
             skipPadding: true,
             returnType: 'bigint',
         });
@@ -38,7 +40,9 @@ test('setting skipPadding to true may not result in greater string token lengths
     const desiredTokenLength = 6;
     let largerTokenFound = false;
     for (let i = 0; i < 10000; i++) {
-        const token = gen(desiredTokenLength, { skipPadding: true });
+        const token = generateSecureToken(desiredTokenLength, {
+            skipPadding: true,
+        });
         largerTokenFound = token.length > desiredTokenLength;
     }
     expect(largerTokenFound).toStrictEqual(false);
@@ -48,7 +52,7 @@ test('setting skipPadding to true may not result in greater number token lengths
     const desiredTokenLength = 6;
     let largerTokenFound = false;
     for (let i = 0; i < 10000; i++) {
-        const token = gen(desiredTokenLength, {
+        const token = generateSecureToken(desiredTokenLength, {
             skipPadding: true,
             returnType: 'number',
         });
@@ -61,7 +65,7 @@ test('setting skipPadding to true may not result in greater bigint token lengths
     const desiredTokenLength = 6;
     let largerTokenFound = false;
     for (let i = 0; i < 10000; i++) {
-        const token = gen(desiredTokenLength, {
+        const token = generateSecureToken(desiredTokenLength, {
             skipPadding: true,
             returnType: 'bigint',
         });
