+++ src/__tests__/warnings.test.ts
@@ -1,5 +1,5 @@
 // @ts-nocheck
-import { generateSecureToken } from '../index';
+import { generateSecureToken } from '../generateSecureToken';
 
 jest.spyOn(console, 'warn').mockImplementation();
