+++ src/library/calculateByteSize.ts
