+++ src/library/constants.ts
@@ -1,4 +1,4 @@
-/**  */
+/** Default size of byte buffer to be generated */
 export const DEFAULT_BYTE_SIZE = 64;
 
 /** 'bigint' return type */
@@ -7,7 +7,7 @@ export const BIGINT = 'bigint';
 /** 'number' return type */
 export const NUMBER = 'number';
 
-/** Legacy 'integer' return type, equivalent to 'number' */
+/** Legacy 'integer' return type, equivalent to using 'number' */
 export const INTEGER = 'integer';
 
 /** 'string' return type */
