+++ src/library/handleOptions.ts
