+++ src/library/validation/validateCustomByteStream.ts
