+++ src/library/validation/validateCustomMemory.ts
