+++ src/library/validation/validateOptions.ts
